require accountId for all authenticated queries

+++ app/routes/board.$id/queries.ts
@@ -2,14 +2,15 @@ import { prisma } from "~/db/prisma";
 
 import { ItemMutation } from "./types";
 
-export function deleteCard(id: string) {
-  return prisma.item.delete({ where: { id } });
+export function deleteCard(id: string, accountId: string) {
+  return prisma.item.delete({ where: { id, Board: { accountId } } });
 }
 
-export async function getBoardData(boardId: number) {
+export async function getBoardData(boardId: number, accountId: string) {
   return prisma.board.findUnique({
     where: {
       id: boardId,
+      accountId: accountId,
     },
     include: {
       items: true,
@@ -18,31 +19,52 @@ export async function getBoardData(boardId: number) {
   });
 }
 
-export async function updateBoardName(boardId: number, name: string) {
+export async function updateBoardName(
+  boardId: number,
+  name: string,
+  accountId: string,
+) {
   return prisma.board.update({
-    where: { id: boardId },
+    where: { id: boardId, accountId: accountId },
     data: { name },
   });
 }
 
-export function upsertItem(mutation: ItemMutation & { boardId: number }) {
+export function upsertItem(
+  mutation: ItemMutation & { boardId: number },
+  accountId: string,
+) {
   return prisma.item.upsert({
-    where: { id: mutation.id },
+    where: {
+      id: mutation.id,
+      Board: {
+        accountId,
+      },
+    },
     create: mutation,
     update: mutation,
   });
 }
 
-export async function updateColumnName(id: string, name: string) {
+export async function updateColumnName(
+  id: string,
+  name: string,
+  accountId: string,
+) {
   return prisma.column.update({
-    where: { id },
+    where: { id, Board: { accountId } },
     data: { name },
   });
 }
 
-export async function createColumn(boardId: number, name: string, id: string) {
+export async function createColumn(
+  boardId: number,
+  name: string,
+  id: string,
+  accountId: string,
+) {
   let columnCount = await prisma.column.count({
-    where: { boardId },
+    where: { boardId, Board: { accountId } },
   });
   return prisma.column.create({
     data: {
